.github/workflows/python_requirements.txt
# Edit Google Google Sheets
gspread
oauth2client
pytz

# Read YAML (.yml or .yaml) file
pyyaml